# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "TPVV",
    "description" : "TPVV Ingeniería Web",
    "contact" : {
      "email" : "lac56@alu.ua.es"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/lac56-alu/TPVV/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://localhost:2000",
    "description" : "Desarrollo servidor basado en TPVV IW"
  }, {
    "url" : "",
    "description" : "Servidor"
  } ],
...
  "paths" : {
...
    "/tpvv/realizarPago" : {
      "post" : {
        "tags" : [ "Pagos" ],
        "summary" : "Se hace una nueva transaccion",
        "description" : "Lo que hacemos es realizar una nueva trasacción vinculada al cliente",
        "operationId" : "realizarPago",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Pago"
        },
        "responses" : {
          "201" : {
            "description" : "Detalles sobre la transaccion",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pagos"
                }
              }
            }
          },
          "400" : {
            "description" : "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_400"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/listaPagos" : {
      "get" : {
        "tags" : [ "Pagos" ],
        "summary" : "Se devuelven todos los pagos realizados por ese usuario",
        "description" : "Lo que hacemos es mediante el token verificar que usuario es el que esta haciendo la peticion y entonces listar todos los pagos que ha realizado a través de nuestra aplicación.",
        "operationId" : "listaPagos",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Lista de los pagos del usuario",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/detallesPago" : {
      "get" : {
        "tags" : [ "Pagos" ],
        "summary" : "Obtiene todos los detalles de una transacción en específico",
        "description" : "Lo que haremos será consultar un pago en concreto. Podremos ver todos los campos que hemos introducido para la transacción así como el estado en el que se encuentra la misma",
        "operationId" : "getDetallesPago",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "idPago",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Detalles sobre la transaccion",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pagos"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          },
          "404" : {
            "description" : "No se encontró el recurso que se pidió",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Pagos" : {
        "required" : [ "referencia", "tarjeta", "total" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1
          },
          "devolucionId" : {
            "type" : "array",
            "description" : "Id de la devolución, si hubiera (en caso contrario -> null)",
            "example" : 22,
            "items" : {
              "type" : "integer"
            }
          },
          "total" : {
            "type" : "number",
            "description" : "Total de la compra",
            "example" : 120.9
          },
          "concepto" : {
            "type" : "string",
            "description" : "Se podrá añadir opcionalmente a la hora de realizar el pago un concepto sobre la compra realizada",
            "example" : "Compra memoria RAM"
          },
          "referencia" : {
            "type" : "string",
            "description" : "Es la referencia que vincula el pago con el recibo creado por el usuario que realiza la compra",
            "example" : "515adfas54"
          },
          "fecha" : {
            "type" : "string",
            "description" : "Fecha del momento en el que se realiza la compra",
            "format" : "DD-MM-YYYY hh:mm:ss",
            "example" : "09-12-2022 09:55:23"
          },
          "estado" : {
            "type" : "string",
            "description" : "Estado de como se encuentra el pago",
            "example" : "DENEGADO",
            "enum" : [ "EN_ESPERA", "ACEPTADO", "DENEGADO, DEVOLUCIÓN" ]
          },
          "detallesEstado" : {
            "type" : "string",
            "description" : "Motivo del error del pago",
            "example" : "TARJETA DENEGADA"
          },
          "tarjeta" : {
            "$ref" : "#/components/schemas/Tarjeta"
          }
        },
        "example" : {
          "detallesEstado" : "TARJETA DENEGADA",
          "fecha" : "09-12-2022 09:55:23",
          "total" : 120.9,
          "estado" : "DENEGADO",
          "concepto" : "Compra memoria RAM",
          "id" : 1,
          "devolucionId" : 22,
          "tarjeta" : {
            "caducidad" : "09/2023",
            "numero" : "984516545251",
            "ccv" : 123,
            "titularTarjeta" : "Luis Alfonso"
          },
          "referencia" : "515adfas54"
        }
      },
      "Tarjeta" : {
        "type" : "object",
        "properties" : {
          "titularTarjeta" : {
            "type" : "string",
            "example" : "Luis Alfonso"
          },
          "numero" : {
            "type" : "string",
            "description" : "Número de la tarjeta de crédito",
            "example" : "984516545251"
          },
          "ccv" : {
            "type" : "number",
            "description" : "Código de seguridad",
            "example" : 123
          },
          "caducidad" : {
            "required" : [ "caducidad", "ccv", "numero", "titularTarjeta" ],
            "type" : "string",
            "description" : "Fecha de caducidad de la tarjeta",
            "format" : "MM/YYYY",
            "example" : "09/2023"
          }
        },
        "example" : {
          "caducidad" : "09/2023",
          "numero" : "984516545251",
          "ccv" : 123,
          "titularTarjeta" : "Luis Alfonso"
        }
      },
...
      "inline_response_400" : {
        "type" : "object",
        "properties" : {
          "typeError" : {
            "type" : "string",
            "example" : "ERROR_PARAMETERS"
          },
          "messageError" : {
            "type" : "string",
            "example" : "Los parámetros introducidos son erroneos."
          }
        }
      },
...
      "inline_response_401" : {
        "type" : "object",
        "properties" : {
          "errorMessage" : {
            "type" : "string",
            "example" : "ERROR_TOKEN"
          }
        }
      },
      "inline_response_200" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Pagos"
            }
          }
        },
        "example" : {
          "data" : [ {
            "detallesEstado" : "TARJETA DENEGADA",
            "fecha" : "09-12-2022 09:55:23",
            "total" : 120.9,
            "estado" : "DENEGADO",
            "concepto" : "Compra memoria RAM",
            "id" : 1,
            "devolucionId" : 22,
            "tarjeta" : {
              "caducidad" : "09/2023",
              "numero" : "984516545251",
              "ccv" : 123,
              "titularTarjeta" : "Luis Alfonso"
            },
            "referencia" : "515adfas54"
          }, {
            "detallesEstado" : "TARJETA DENEGADA",
            "fecha" : "09-12-2022 09:55:23",
            "total" : 120.9,
            "estado" : "DENEGADO",
            "concepto" : "Compra memoria RAM",
            "id" : 1,
            "devolucionId" : 22,
            "tarjeta" : {
              "caducidad" : "09/2023",
              "numero" : "984516545251",
              "ccv" : 123,
              "titularTarjeta" : "Luis Alfonso"
            },
            "referencia" : "515adfas54"
          } ]
        }
      },
      "inline_response_404" : {
        "type" : "object",
        "properties" : {
          "errorMessage" : {
            "type" : "string",
            "example" : "ERROR_RECURSO_NO_ENCONTRADO"
          }
        }
      },
...
    },
...
    "requestBodies" : {
...
      "Pago" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Pagos"
            }
          }
        },
        "required" : true
      },
...
    }
  }
}